requirements.txt
numpy==1.24.3
Pillow==9.4.0
Pillow==10.2.0
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
wandb==0.15.8
